refactor: Phase C2 - extract Admin AI config and schedule sections
AdminPage (1386→336 lines):
- AdminAiConfigSection: provider toggle, API keys, model select, prompts, test connections
- AvailabilityScheduleSection: settings form, run history table, expandable details, pagination
- ValuationScheduleSection: settings form, trigger/cancel, run history, polling, pagination

Schedule components are self-contained with their own API calls and state.
AI config receives shared settings via props from parent's useAdminConfig composable.
vue-tsc passes clean.
